:truck: Move collection tests to own subfolder
