+++ collections/tests/filter_entries_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../testing/asserts.ts";
-import { filterEntries } from "./filter_entries.ts";
-import { Predicate } from "./types.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { filterEntries } from "../filter_entries.ts";
+import { Predicate } from "../types.ts";
 
 function filterEntriesTest<T>(
   input: [Record<string, T>, Predicate<[string, T]>],
