+++ collections/tests/find_last_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../testing/asserts.ts";
-import { findLast } from "./find_last.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { findLast } from "../find_last.ts";
 
 function findLastTest<I>(
   input: [Array<I>, (element: I) => boolean],
