+++ collections/tests/intersect_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../testing/asserts.ts";
-import { intersect } from "./intersect.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { intersect } from "../intersect.ts";
 
 function intersectTest<I>(
   input: Array<Array<I>>,
