+++ collections/tests/map_values_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../testing/asserts.ts";
-import { mapValues } from "./map_values.ts";
-import { Selector } from "./types.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { mapValues } from "../map_values.ts";
+import { Selector } from "../types.ts";
 
 function mapValuesTest<T, O>(
   input: [Record<string, T>, Selector<T, O>],
