+++ collections/tests/partition_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../testing/asserts.ts";
-import { partition } from "./partition.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { partition } from "../partition.ts";
 
 function partitionTest<I>(
   input: [Array<I>, (element: I) => boolean],
