+++ collections/tests/sort_by_test.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 
-import { assertEquals } from "../testing/asserts.ts";
-import { sortBy } from "./sort_by.ts";
-import { Selector } from "./types.ts";
+import { assertEquals } from "../../testing/asserts.ts";
+import { sortBy } from "../sort_by.ts";
+import { Selector } from "../types.ts";
 
 function sortByTest<T>(
   input: [
